revert the ComposeWindow.web.kt split
